Move default implementations so the Launcher's serializer can find them
