+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/protocol/events/DefaultAnnotationProcessorResult.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.tasks.compile.tooling;
+package org.gradle.tooling.internal.protocol.events;
 
 import org.gradle.tooling.internal.protocol.events.InternalJavaCompileTaskSuccessResult.InternalAnnotationProcessorResult;
